ConfigurationTest is now divided into unit one and automation one
